Correctly set local options in vimrc
It has taken me a *long* time to internalize this, but calling
:setglobal on a local option sets the default for new buffers, but does
not affect the very first buffer that already exists when the vimrc is
loaded.  This is distinct from "global-local" options, which delegate to
the global option when the local option is an empty string.

This also gives up on 'cmdheight', which is documented as global but
appears to have become tab specific at some point.  Keeping 'cmdheight'
at 1 keeps me diligent about minimizing "press ENTER" prompts in my
vimrc, so let's do that rather than attempt to fix it.